Add files via upload
This pictures shows why, when using less than 5000 services on the same host, it makes sense using a random port choice. This way the programmer does not need to care about which ones are the used ports but just launches the services. When reaching over the 10000 used ports, choosing the ports randomly it's not worth it anymore: the possibility of using an already used port iot's about 17.5%, which is to big to be affordable.